Uspjenso napravljen location DDM
