+++ Front/src/typings/DDM-DropDownMenu/DDM.tsx
@@ -1,0 +1,5 @@
+export interface LocationDDMProps {
+  options: { [key: string]: string[] };
+  description: string;
+  placeHolder: string;
+}
