+++ Front/src/typings/locationCategorys/LocationCategroys.tsx
@@ -1,0 +1,4 @@
+export type Category =
+  | 'Cities (including airports)'
+  | 'Airports'
+  | 'Train stations';
